chore: code folder structure
